Logic fix: do not pretend that a parsing error means no arguments
Getopt::Long::GetOptions returns false if the list of options passed to
it was incorrect, e.g. undef values in the list, or more complicated
errors to do with callbacks. If this happens you'll get an error message
to the console, and the code should quit.

The code appeared to be thinking that it would return false if e.g. no
arguments at all had been passed, which is not the case. If you do that,
you get

  error: explicit output directory (-p) required

instead, which is thrown by xgettext-perl a fair bit later when it
unpacks its arguments.